Fix call forwarding indication.
The actual call reference is the Connection object, which is passed
around the call objects. Previously, it might have happened that a call
stayed in the forwarded call list because the connection owner at
disconnection time wasn't the same as at connection time. Avoid this by
managing the connections instead of the calls in the forwarded calls
list.

Change-Id: I02491773f8a1b154974b83f984e5aa16f535e35f